feat: Upgrade cardano-common-explorer to version 1.0.7

+++ src/main/java/com/cardano/explorer/model/response/stake/StakeAnalyticRewardResponse.java
@@ -7,6 +7,6 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public class StakeAnalyticRewardResponse {
-  private Long epoch;
+  private Integer epoch;
   private BigInteger value;
 }
